+++ recipes/default.rb
@@ -17,11 +17,6 @@
 # limitations under the License.
 #
 
-include_recipe "prestashop::database" if node['prestashop']['install_db'] == true
 include_recipe "prestashop::php5_mcrypt" if node['prestashop']['with_php5_mcrypt'] == true
 include_recipe "prestashop::php5_imap" if node['prestashop']['need_imap_for_service_client'] == true
-include_recipe "prestashop::install"
-include_recipe "prestashop::apache_vhost"
-include_recipe "prestashop::other_modules"
-include_recipe "prestashop::piwik" if node['prestashop']['other_modules']['get_piwik'] == true
 include_recipe "prestashop::presh" if node['prestashop']['presh']['enabled'] == true