+++ recipes/presh.rb
@@ -17,41 +17,28 @@
 # limitations under the License.
 #
 
-install_dir = node.default['prestashop']['presh']['install_dir_base'] +
-  node.default['prestashop']['presh']['install_dir_suffix'] + '/'
-revision = node.default['prestashop']['presh']['revision']
+presh_dir = (
+  node['prestashop']['presh']['install_dir_base'] +
+  node['prestashop']['presh']['install_dir_suffix']
+)
+revision = node['prestashop']['presh']['revision']
 
 bash 'get-presh-from-github' do
-  download_url = node.default['prestashop']['presh']['base_url'] +
-    node.default['prestashop']['presh']['revision'] + '.zip'
+  download_url = node['prestashop']['presh']['base_url'] +
+    node['prestashop']['presh']['revision'] + '.zip'
   downloaded_file_name = 'presh-' + revision + '.zip'
   # downloaded_directory will contain our target directory (e.g presh-master)
-  downloaded_directory = 'presh-' + node.default['prestashop']['presh']['revision']
+  downloaded_directory = 'presh-' + node['prestashop']['presh']['revision']
   code <<-EOH
     wget --continue --quiet #{download_url} --output-document=#{downloaded_file_name}
     unzip -o #{downloaded_file_name}
-    rm -rf #{install_dir}
-    mv #{downloaded_directory} #{install_dir}
-    chmod 755 #{install_dir}presh
+    rm -rf #{presh_dir}
+    mv #{downloaded_directory} #{presh_dir}
+    chmod 755 #{::File.join(presh_dir, 'presh')}
     rm -f #{downloaded_file_name}
   EOH
   only_if {
-    keep_updating = node.default['prestashop']['presh']['keep_updating']
-    Dir.exists?(install_dir) == false || (keep_updating == true && revision == 'master')
+    keep_updating = node['prestashop']['presh']['keep_updating']
+    Dir.exists?(presh_dir) == false || (keep_updating == true && revision == 'master')
   }
-end
-
-bash 'apply-presh-commands' do
-  user node['prestashop']['dir_owner']
-  group node['prestashop']['dir_group']
-  presh_exec = install_dir + 'presh'
-  bash_commands = ''
-  node.default['prestashop']['presh']['commands'].each do |command|
-    bash_commands << <<-EOH
-      #{presh_exec} #{command['name']} #{command['params'] || ''}
-    EOH
-  end
-  cwd node.default['prestashop']['base_dir']
-  code bash_commands
-  only_if {!node.default['prestashop']['presh']['commands'].empty?}
 end